+++ future_context2/lib/src/impl/future_context_impl.dart
@@ -175,7 +175,7 @@ class FutureContextImpl implements FutureContext {
 
     var done = false;
     T2? result;
-    Exception? exception;
+    Object? error;
     StackTrace? stackTrace;
 
     // 非同期処理を開始する
@@ -186,8 +186,8 @@ class FutureContextImpl implements FutureContext {
         // 実行前に、実行可能状態まで待機する.
         await resume();
         result = await block(this);
-      } on Exception catch (e, s) {
-        exception = e;
+      } on Object catch (e, s) {
+        error = e;
         stackTrace = s;
       } finally {
         done = true;
@@ -199,15 +199,16 @@ class FutureContextImpl implements FutureContext {
     // キャンセルが発生していたらキャンセル例外を投げる.
     // 完了したら、結果を返却する.
     await for (final _ in _controller.notifyStream) {
-      if (exception != null) {
+      if (error != null) {
         // 発生した例外をチェックする.
-        if (exception is CancellationException) {
+        if (error is CancellationException) {
           throw CancellationException(
-            '${toString()} is canceled. caused by exception: $exception, stackTrace: $stackTrace',
+            '${toString()} is canceled. caused by exception: $error, stackTrace: $stackTrace',
           );
         } else {
           // 発生した例外をそのまま投げる
-          throw exception!;
+          // ignore: only_throw_errors
+          throw error!;
         }
       } else if (isCanceled) {
         // その他、キャンセルが発生していたらキャンセルを投げる.